roda_core/requirements.txt
Django==2.0
django-extensions==1.9.8
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
pytz==2017.3
